requirements.txt

flask==2.3.3
flask-cors==4.0.0
langchain-core==0.3.15
langchain-groq==0.2.0
requests==2.31.0
python-dotenv==1.0.0